Resolve conflicts addVendorModal and navbar
