+++ src/App.scss
@@ -53,7 +53,7 @@ nav {
   left: 0;
   width: 100%;
   background-color: #f2f2f2;
-  z-index: 4;
+  z-index: 8;
 }
 
 .backdrop {
@@ -62,6 +62,6 @@ nav {
   left: 0;
   width: 100%;
   height: 100vh;
-  z-index: 10;
+  z-index: 7;
   background-color: rgba(0, 0, 0, 0.75);
 }
