+++ src/styles/components/Spinner.scss
@@ -1,0 +1,25 @@
+.spin {
+    position: fixed;
+    height: 50px;
+    width: 50px;
+    top: 50%;
+    left: 50%;
+
+    border: 5px solid rgba(255, 255, 255, 0.3);
+    border-radius: 50%;
+    border-top-color: #003d4f;
+    animation: spin 1s ease-in-out infinite;
+    -webkit-animation: spin 1s ease-in-out infinite;
+    z-index: 80;
+  }
+
+  @keyframes spin {
+    to {
+      -webkit-transform: rotate(360deg);
+    }
+  }
+  @-webkit-keyframes spin {
+    to {
+      -webkit-transform: rotate(360deg);
+    }
+  }
